create useSearch
